ssbuild 9.15

+++ mobile/src/main/java/cn/garymb/ygomobile/ui/home/ImageUpdater.java
@@ -1,7 +1,9 @@
 package cn.garymb.ygomobile.ui.home;
 
+import android.app.Activity;
 import android.content.Context;
 import android.content.DialogInterface;
+import android.util.Log;
 import android.widget.Toast;
 
 import java.io.File;
@@ -20,12 +22,14 @@ import java.util.zip.ZipFile;
 
 import cn.garymb.ygomobile.AppsSettings;
 import cn.garymb.ygomobile.Constants;
-import cn.garymb.ygomobile.bean.CardInfo;
 import cn.garymb.ygomobile.loader.CardLoader;
 import cn.garymb.ygomobile.lite.R;
+import cn.garymb.ygomobile.ui.activities.BaseActivity;
 import cn.garymb.ygomobile.ui.plus.DialogPlus;
 import cn.garymb.ygomobile.ui.plus.VUiKit;
+import cn.garymb.ygomobile.utils.FileUtils;
 import cn.garymb.ygomobile.utils.IOUtils;
+import ocgcore.data.Card;
 import ocgcore.enums.CardType;
 
 /**
@@ -33,7 +37,7 @@ import ocgcore.enums.CardType;
  */
 
 public class ImageUpdater implements DialogInterface.OnCancelListener {
-    private Context mContext;
+    private BaseActivity mContext;
     private CardLoader mCardLoader;
     private final static int SubThreads = 4;
     private int mDownloading = 0;
@@ -50,7 +54,7 @@ public class ImageUpdater implements DialogInterface.OnCancelListener {
 
     File mPicsPath;
 
-    public ImageUpdater(Context context) {
+    public ImageUpdater(BaseActivity context) {
         mContext = context;
         mCardLoader = new CardLoader(context);
         mPicsPath = new File(AppsSettings.get().getResourcePath(), Constants.CORE_IMAGE_PATH);
@@ -187,16 +191,21 @@ public class ImageUpdater implements DialogInterface.OnCancelListener {
             }
             if (needNext) {
                 if (existImage()) {
+
                 } else {
-                    if (download(item.url, tmpFile) && tmpFile.exists()) {
+                    if (download(item.url, tmpFile)) {
                         File file = new File(item.file);
-                        if (!file.exists()) {
+                        if (tmpFile.exists() && !file.exists()) {
                             tmpFile.renameTo(file);
+                            Log.d("kk", "download ok:" + item.url + " ->" + item.file);
+                        } else {
+                            Log.e("kk", "download fail:" + item.url + " ->" + item.file);
                         }
                     } else {
                         synchronized (mCardStatus) {
                             mError++;
                         }
+                        Log.e("kk", "download error:" + item.url + " ->" + item.file);
                     }
                 }
                 synchronized (mCardStatus) {
@@ -248,10 +257,20 @@ public class ImageUpdater implements DialogInterface.OnCancelListener {
         HttpURLConnection mConnection = null;
         boolean ok = false;
         try {
+            if (file.exists()) {
+                file.delete();
+            } else {
+                File dir = file.getParentFile();
+                if (!dir.exists()) {
+                    dir.mkdirs();
+                }
+            }
+            file.createNewFile();
             mConnection = (HttpURLConnection) new URL(url).openConnection();
             mConnection.setConnectTimeout(30 * 1000);
             mConnection.setReadTimeout(15 * 1000);
-            mConnection.setUseCaches(false);
+            mConnection.setRequestProperty("User-Agent",
+                    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/52.0.2743.116 Safari/537.36 Edge/15.15063");
             mConnection.connect();
             if (mConnection.getResponseCode() == 200) {
                 inputStream = mConnection.getInputStream();
@@ -263,9 +282,14 @@ public class ImageUpdater implements DialogInterface.OnCancelListener {
                 }
                 outputStream.flush();
                 ok = true;
+            } else {
+                Log.w("kk", "download:" + mConnection.getResponseCode() + ":url=" + url);
             }
         } catch (IOException e) {
-
+            Log.w("kk", "download:" + url, e);
+            if (file.exists()) {
+                file.delete();
+            }
         } finally {
             IOUtils.close(inputStream);
             IOUtils.close(outputStream);
@@ -304,9 +328,9 @@ public class ImageUpdater implements DialogInterface.OnCancelListener {
         }
         VUiKit.post(() -> {
             if (mError == 0) {
-                Toast.makeText(mContext, R.string.downloading_images_ok, Toast.LENGTH_SHORT).show();
+                mContext.showToast(R.string.downloading_images_ok, Toast.LENGTH_SHORT);
             } else {
-                Toast.makeText(mContext, mContext.getString(R.string.download_image_error, mError), Toast.LENGTH_SHORT).show();
+                mContext.showToast(mContext.getString(R.string.download_image_error, mError), Toast.LENGTH_SHORT);
             }
         });
     }
@@ -315,15 +339,15 @@ public class ImageUpdater implements DialogInterface.OnCancelListener {
         if (!mCardLoader.isOpen()) {
             mCardLoader.openDb();
         }
-        Map<Long, Long> cards = mCardLoader.readAllCardCodes();
+        Map<Long, Card> cards = mCardLoader.readAllCardCodes();
         mCardStatus.clear();
         mPicsPath = new File(AppsSettings.get().getResourcePath(), Constants.CORE_IMAGE_PATH);
         File picsPath = mPicsPath;
         File fieldPath = new File(mPicsPath, Constants.CORE_IMAGE_FIELD_PATH);
         IOUtils.createNoMedia(picsPath.getAbsolutePath());
         IOUtils.createNoMedia(fieldPath.getAbsolutePath());
-        for (Map.Entry<Long, Long> e : cards.entrySet()) {
-            if (CardInfo.isType(e.getValue(), CardType.Field)) {
+        for (Map.Entry<Long, Card> e : cards.entrySet()) {
+            if (Card.isType(e.getValue().Type, CardType.Field)) {
                 String png = new File(fieldPath, e.getKey() + Constants.IMAGE_FIELD_URL_EX).getAbsolutePath();
                 String pngUrl = String.format(Constants.IMAGE_FIELD_URL, e.getKey() + "");
                 mCardStatus.add(new Item(pngUrl, png, e.getKey(), true));
